refactor(pages): move static content to config.ts in studio, home and appointment pages
