+++ frontend/src/app/features/appointment/appointment-page/appointment-page.component.html
@@ -1,7 +1,7 @@
 <div class="bg-granite py-20 px-6">
   <div class="max-w-xl mx-auto">
     <div class="bg-white p-6 rounded-lg shadow mt-10">
-      <h1 class="text-center text-4xl font-semibold mb-6 text-balance">Estás a un paso de pedir cita en FINE LINE</h1>
+      <h1 class="text-center text-4xl font-semibold mb-6 text-balance">{{appointmentData.formSection.mainTitle}}</h1>
       <app-appointment-form></app-appointment-form>
     </div>
   </div>
