+++ frontend/src/app/features/appointment/appointment-page/appointment-page.component.ts
@@ -1,5 +1,6 @@
 import { Component } from '@angular/core';
 import { AppointmentFormComponent } from '../appointment-form/appointment-form.component';
+import { appointmentPageData } from './appointment-page.config';
 
 @Component({
   selector: 'app-appointment-page',
@@ -7,4 +8,6 @@ import { AppointmentFormComponent } from '../appointment-form/appointment-form.c
   templateUrl: './appointment-page.component.html',
   styleUrl: './appointment-page.component.scss',
 })
-export class AppointmentPageComponent {}
+export class AppointmentPageComponent {
+  appointmentData = appointmentPageData;
+}
