+++ frontend/src/app/features/appointment/appointment-page/appointment-page.config.ts
@@ -1,0 +1,5 @@
+export const appointmentPageData = {
+  formSection: {
+    mainTitle: 'Estás a un paso de pedir cita en FINE LINE',
+  },
+};
