+++ frontend/src/app/features/studio/studio-page/studio-page.config.ts
@@ -19,17 +19,22 @@ export const studioPageData = {
     imageStyle: 'square',
   } as const,
 
-  microrealismImages: [
-    'assets/images/tattoos/tattoo-bird.webp',
-    'assets/images/tattoos/tattoo-face.webp',
-    'assets/images/tattoos/tattoo-hands.webp',
-  ],
-
-  geometricsImages: [
-    'assets/images/tattoos/tattoo-plant.webp',
-    'assets/images/tattoos/tattoo-sculture.webp',
-    'assets/images/tattoos/tattoo-arm.webp',
-  ],
+  microrealismSection: {
+    mainTitle: 'Tatuajes microrealistas',
+    microrealismImages: [
+      'assets/images/tattoos/tattoo-bird.webp',
+      'assets/images/tattoos/tattoo-face.webp',
+      'assets/images/tattoos/tattoo-hands.webp',
+    ],
+  },
+  geometricsSection: {
+    mainTitle: 'Tatuajes geométricos',
+    geometricsImages: [
+      'assets/images/tattoos/tattoo-plant.webp',
+      'assets/images/tattoos/tattoo-sculture.webp',
+      'assets/images/tattoos/tattoo-arm.webp',
+    ],
+  },
 
   micropigmentationSection: {
     title: 'Micropigmentación',
